+++ ui/actions/findings/findings.ts
@@ -92,7 +92,7 @@ export const getMetadataInfo = async ({
 
   Object.entries(filters).forEach(([key, value]) => {
     // Define filters to exclude
-    const excludedFilters = ["region__in", "service__in", "resource_type__in"];
+    const excludedFilters = ["region__in", "service__in", "resource_type__in", "categories__in"];
     if (
       key !== "filter[search]" &&
       !excludedFilters.some((filter) => key.includes(filter))
@@ -127,7 +127,7 @@ export const getLatestMetadataInfo = async ({
 
   Object.entries(filters).forEach(([key, value]) => {
     // Define filters to exclude
-    const excludedFilters = ["region__in", "service__in", "resource_type__in"];
+    const excludedFilters = ["region__in", "service__in", "resource_type__in", "categories__in"];
     if (
       key !== "filter[search]" &&
       !excludedFilters.some((filter) => key.includes(filter))