+++ prowler/CHANGELOG.md
@@ -41,6 +41,7 @@ All notable changes to the **Prowler SDK** are documented in this file.
 ## [v5.14.0] (Prowler v5.14.0)
 
 ### Added
+- Category filter for CLI Dashboard and Prowler App UI [(#6646)](https://github.com/prowler-cloud/prowler/pull/6646)
 - GitHub provider check `organization_default_repository_permission_strict` [(#8785)](https://github.com/prowler-cloud/prowler/pull/8785)
 - Add OCI mapping to scan and check classes [(#8927)](https://github.com/prowler-cloud/prowler/pull/8927)
 - `codepipeline_project_repo_private` check for AWS provider [(#5915)](https://github.com/prowler-cloud/prowler/pull/5915)